change datatype of amount in payment table

+++ model/dekshina.js
@@ -11,7 +11,7 @@ const dekshina = sequelize.define('dekshina', {
     razorpay_payment_id: { type: DataTypes.STRING},
     razorpay_signature: { type: DataTypes.STRING}, 
     UId:{ type: DataTypes.INTEGER}  ,
-    amount : { type : DataTypes.INTEGER},
+    amount : { type : DataTypes.DOUBLE},
     payment_date :{ type:DataTypes.STRING},
     payment_time:{ type: DataTypes.STRING},
     dekshina_payment_status:{ type: DataTypes.BOOLEAN}
